fix(schools): corrigido funcionalidades, reduzido instrucoes e adicionado funcionalidades

+++ app/Http/Controllers/EscolaController.php
@@ -2,31 +2,26 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Escola;
 use App\Models\Municipio;
-use Illuminate\Http\RedirectResponse;
-use Illuminate\View\View;
 use App\Http\Requests\StoreEscolaRequest;
 use App\Http\Requests\UpdateEscolaRequest;
-use App\Http\Resources\EscolaResource;
-use App\Models\Escola;
-use Illuminate\Http\Request;
-use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
-use Illuminate\Support\Facades\Auth; 
+use Illuminate\Http\RedirectResponse;
+use Illuminate\View\View;
+
 class EscolaController extends Controller
 {
-
     public function index(): View
     {
-        $usuario = Auth::user();
-        $escolasQuery = Escola::query()->with('municipio');
-        if ($usuario->tipo_usuario !== 'administrador' && $usuario->id_escola) {
-            $escolasQuery->where('id_escola', $usuario->id_escola);
-        }
-        
-        $escolas = $escolasQuery->orderBy('nome')->get();
-        
+        $escolas = Escola::with([
+                            'municipio', 
+                            'usuarios' => fn($query) => $query->where('tipo_usuario', 'diretor')
+                        ])
+                        ->orderBy('nome')
+                        ->get();
+                        
         $municipios = Municipio::orderBy('nome')->get();
+        
         return view('schools.index', compact('escolas', 'municipios'));
     }
 
@@ -36,41 +31,29 @@ class EscolaController extends Controller
         return redirect()->route('escolas.index')->with('success', 'Escola adicionada com sucesso!');
     }
 
-    public function show(Escola $escola): EscolaResource
-    {
-        $this->authorize('view', $escola); 
-        
-        $escola->load(['municipio', 'diretor', 'turmas', 'usuarios']);
-        
-        return new EscolaResource($escola);
-    }
-
     public function edit(Escola $escola): View
     {
-         // Aplicar Policy
-        // $this->authorize('update', $escola); 
-        
         $municipios = Municipio::orderBy('nome')->get();
         return view('schools.edit', compact('escola', 'municipios'));
     }
 
     public function update(UpdateEscolaRequest $request, Escola $escola): RedirectResponse
     {
-         // Aplicar Policy
-        // $this->authorize('update', $escola); 
-
         $escola->update($request->validated());
-
         return redirect()->route('escolas.index')->with('success', 'Escola atualizada com sucesso!');
     }
 
     public function destroy(Escola $escola): RedirectResponse
     {
-         // Aplicar Policy (só admin)
-        // $this->authorize('delete', $escola); 
-        
-        $escola->delete();
+        if ($escola->turmas()->exists()) {
+            return redirect()->route('escolas.index')->with('error', 'Não é possível excluir esta escola, pois ela já possui turmas cadastradas.');
+        }
 
+         if ($escola->usuarios()->exists()) {
+            return redirect()->route('escolas.index')->with('error', 'Não é possível excluir esta escola, pois ela possui usuários (diretores ou professores) associados.');
+        }
+
+        $escola->delete();
         return redirect()->route('escolas.index')->with('success', 'Escola excluída com sucesso!');
     }
 }
